requirements.txt
Django~=5.0.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-dotenv==0.19.0
requests~=2.32.3
beautifulsoup4~=4.12.3